Default to MATSBAIJ for all operators [partial]
With petsc-3.2, one would call DMGetMatrix(dm,mtype,&mat) and mtype
would be ignored if the user passed -da_mat_type.  There was no way for
the caller to override the command-line option.  This is inconsistent
with the rest of PETSc, so the mtype argument was removed and
DMSetMatType was added.  But pTatin wants the commannd line to be
preferred, so set the default type at creation and then call
DMSetFromOptions.