Re-factor manner in which coefficient projections are defined. The option   -coefficient_projection_type now admits   - Q1 projection <default>   - P0 with arithmetic averaging over a Q2 element   - P0 with harmonic averaging over a Q2 element   - P0 with geometric averaging  over a Q2 element
-coefficient_projection_type 01 (or 1) -> Q1
-coefficient_projection_type 00 (or 0) -> P0 (arith)
-coefficient_projection_type 10 -> P0 (harm)
-coefficient_projection_type 20 -> P0 (geom)

The P0 family compute an element value of the viscosity
and apply it as a constant on each quadrature point
within a given Q2 element.

Changes:
* Two major classes of projections are distinguished when
evaluating coarse level viscosity fields.
  Class 1: Fine level projections which result in values on
the quadrature points. In this case, coarse level viscosity
fields are computed by first projecting the quadrature
point values onto a Q1 space on the fine level, and this
Q1 viscosity value is restricted through the hierarchy.
Note

  Class 2: Fine level projections which result in a 
Q1 nodal viscosity field. In this case, all that is
required is that the nodal viscosity is restricted
through the hierarchy.

[NOTE] Neither class 1 or class 2 modify the quadrature point
viscosity field defined on the finest level.