Makefile: use ln -sf to force link, instead of ln -sF
The -F option is for forcing linking, if the target is a directory.
Some implementations of ln (for instance, the one on my current OS X laptop)
take -F to imply -f (remove target if it exists, whether or not it's a
directory), but not all. This caused a link error when building on an
Ubuntu machine with a different ln . -f appears to work on both machines.