WIP (testable): Add NodeRepart MF A11 operator
This is WIP to add a new A11 matrix-free implementation.
It aims to take advantage of different resources available
on different ranks of a (shared memory) node. In particular,
we have nodes on Piz Daint which have 12 CPU cores attached
to a UMA memory domain, and one GPU. We'd like to be able to
shift some of the elements to be processed to rank 0 on
a communicator defined on each node, process them with the GPU,
and then return the results to the individual per-core ranks.

This proceeds by transferring portions of the arrays to be processed,
namely those defining input velocity values, element coordinates,
element-->node maps (renumbered), and coefficient information.
These arrays are processed as usual (on a per-element basis),
and then returned to their original ranks.

This is in some ways less efficient than directly redistributing data,
but it has the advantage that tasks which must be performed on the
CPU can be peformed wrt the original DMDA partition, and will
thus be balanced.

This commit adds logic which mimicks the "tensor" implementation,
with the change that (inspired by the CUDA implementation),
per-quadarature-point information is precomputed as
"guassdata_w" before redistribution, instead of distributing
the weights and full coefficient structs.

A new logging stage is added, called MatMultMFA11_sub.
This wraps the rank-local part of the work, not including
auxiliary communication and copying. Thus, the "ratio" of times
for this stage as seen with -log_view should give and idea
of the imbalance.

A flag -subrepart_size (default value 12) controls how
the subdomains are distributed. Currently, contiguous chunks of
ranks are selected.

The flag -subrepart_frac controls how much of each non-root rank,
with respect to the subdomain communicator, should be sent to
the root rank of each subcommunicator. Thus, if you have 12 ranks
per subdomain, and you believe that rank 0 can process as much
data as the 11 other ranks combined, you should set this value to 0.5.
If you believe that rank 0 can process data half as fast as the other 11
ranks combined, you should set this value to 0.33.

The next steps are to update to using the AVX implementation (similar),
and then to introduce (optionally) the CUDA implementation on sub-rank 0.

Test with e.g.
````bash
export PMPI=$PETSC_DIR/$PETSC_ARCH/bin/mpiexec
$PMPI -n 3 $PETSC_ARCH/bin/test_stokes_operators.app -compare_operators -a11_op subrepart -mx 1 -my 1 -mz 3 -ptatin_model viscous_sinker -model_viscous_sinker_eta1 10
$PMPI -n 7 $PETSC_ARCH/bin/test_stokes_operators.app -compare_operators -a11_op subrepart -mx 3 -my 7 -mz 11 -ptatin_model viscous_sinker -model_viscous_sinker_eta1 10
$PMPI -n 16 $PETSC_ARCH/bin/test_stokes_operators.app -compare_operators -a11_op subrepart -mx 5 -my 7 -mz 14 -ptatin_model viscous_sinker -model_viscous_sinker_eta1 10
$PMPI -n 64 $PETSC_ARCH/bin/test_stokes_operators.app -compare_operators -a11_op subrepart -mx 16 -my 16 -mz 16 -ptatin_model viscous_sinker -model_viscous_sinker_eta1 10
````

````bash
$PMPI -n 12 $PETSC_ARCH/bin/test_stokes_operators.app -compare_operators -a11_op subrepart -mx 12 -my 12 -mz 12 -ptatin_model viscous_sinker -model_viscous_sinker_eta1 10 -log_view -subrepart_frac 0.25
````