WIP: pyTestHarness test groups
Change the test naming scheme to use dots instead of path separators.

Add some test targets to the makefile.
With the help of python script, define some groups of tests.
The default "make test" does the "smoke" tests, which are supposed to be fast.
"make testall" runs all the tests.
"make testclassic" runs the legacy tests (currently these are the only tests defined)