material_point_utils.c, model_utils.c: add initialization to resolve annoying GCC warning
GCC throws -Wmaybe-uninitialized warnings here (with GCC 6.3.0, only with -O1 and above, which seems like a bug). This might be something to do with the fact that in theory the value of an enum could be any integer value.

I don't like having to do totally unnecessary initialization, but this is preferable to suppressing entire categories of warnings. It'd also be possible to insert #pragma's to have GCC ignore these specific warnings.