Manual: update installation section, update to allow compilation
There were some conflicts related to the scrartcl class which were
preventing compilation, so some changes were made (\it --> \textit, \sc--> \textsc, and then a blanket hack taken from googling).

Some updates are added as noticed to the installation sections. 
The location of the version file is preemptively changed to be in include/.