SubRepart Operator: Add more diagnostics and warnings
In preparation for a potential optimization, compute what part of the
range of dof referenced on a particular rank are referenced by the
elements there. Since we expect this to be high,
1. Our renumbering scheme may be overcomplex and wasteful - it would be
better to use a simpler numbering (just offset) and a small number of
redundant entries
2. We might be able to intervene earlier in the operator application,
scattering directly to/from global vectors to the repartitioned
structures.