SubRepart : using contiguous node ranges
This simplifies the transfers to and from the shared memory arrays
by using a contiguous range of node numbers on each rank_sub > 0.
This should speed the transfer process, at the cost of shuffling around
some redundant node data that isn't used.

It should also simplify and speed the setup process for this operator
(which doesn't really matter so much, as we typically set these
operators up once and apply them many times in the MG hierarchy).