Prefix installs : allow, installing into root dir
This mimicks the structure of $PETSC_DIR, installing into a $PETSC_ARCH
directory when $PETSC_ARCH is defined, and directly in to the root
directory when $PETSC_ACCH is not defined.

The test scripts are updated to interpret an non-existent PETSC_ARCH
as PETSC_ARCH='' . This is a convenience when working with paths.