Makefile: use new -w option for pyTestHarness (0.2.8) for separate configs
Store pth.conf in the install directory and use it for the tests.

This allows, most importantly, to use the correct mpiexec for each
different PETSc install.

For prefix installs, as with other files, pth.conf will end up in the
root directory.

Note that this does not remove this configuration file with "make clean"
(but maybe it should)